split cc_mesh class into cc_mesh and cc_contour
